fix: Segfaults on null query
The current design assumes that the QStateMachine works synchronous and that the
signal connections are direct, which should have allowed to assume that the
query is never null when calling the relevant functions _by contract_.

However QStateMachine has its own event queue and events are processed
asynchronously which leads to the query potentially beeing null when the signal
guards are called.

A proper fix requires a redesign of the frontend statemachine usage which is some
more work than just this quick fix. Checking for null may prevent segfaults but
probably does not ensure a proper determinsitic state of the machine.